Stripe payment was successfully reached to the Stripe server. However, there are a few bugs.
The User's cart is not cleared after the checkout.
The order post method needs to be changed, do not store credit card info to the database, remove the payment object.
The payment table which was generated by the order's table should be deleted.
The validations aren't added to the credit card field.
If the userId is 0 from the session storage, please manually refresh the page.